VLV-02 | Possible integer overflow and dead loop

+++ contracts/Vaults/VaultLocked.sol
@@ -264,7 +264,7 @@ contract VaultLocked is IDistributable, Ownable, ReentrancyGuard, DepositoryRest
     }
 
     function _deleteUser(address _account) private {
-        for (uint8 i = 0; i < users.length; i++) {
+        for (uint i = 0; i < users.length; i++) {
             if (users[i] == _account) {
                 delete users[i];
             }
